change hyphen to em-dash and backup to back up
The em-dash should be used to connect the two parts of this sentence, rather than a hyphen. In addition, backup is a noun, and should be replaced with the verb.

Although these changes may seem trivial, this greeting is present at every start-up, and plays an important role in the wallet appearing professional.