Speed up travis builds: enable ccache, cache cargo and reduce test timeouts (#957)
* Do not hang trying to upload to docker in unofficial repos

docker push hung asking for a password if the password environment
variable was not set. With this change it fails immediately.

* Enable ccache on linux and osx

* Cache homebrew downloads

This does not avoid the compilation of the homebrew packages but stores
the downloaded sources, speeding up the installation process.

* Enable cargo caching

* Cache load-tester cargo binaries

* Reduce default timeout per test suite to 120 seconds

One of the tests hangs in the CI system, reducing the timeout to have
quicker feedback. It can be re-adjusted in the future if 120 is not
enough for a successful run.