Ensure the winner of an election is always in confirmation or confirmed
Currently a second election may be started for a confirming block, which is not intended but also not critical.

Also cleans up active_transactions tests and, unrelated, RPC pending loop non-functional change, only because this issue was discovered through that RPC.

An intermittent failure in confirmation_height.pending_observer_callbacks is fixed by waiting for cementing.