Publish prototype (#2468)
This is commit is based on a contribution from @Srayman on a recommendation to reduce traffic to principle representatives. The substantive changes are:
- Principle representatives do not relay votes
- Principle representatives are not the target of vote relaying
- Each principle representative broadcasts its votes to all other principle representative and additional non-principle representatives.

An analysis of this methodology as well as other suggestions were posted https://medium.com/nanocurrency/proposal-for-nano-node-network-optimizations-21003e79cdba . 

Testing on beta shows upwards of a 50% reduction in traffic to principle representatives.