Incremental options for ws confirmation subscription
Possibility to add or remove accounts in an existing subscription. This is useful for external wallets that can't use the all_local_accounts flag.

Will also note in the documentation that it should not be used for a large number of accounts since it's a vector search for every confirmation.