Delay voting for non-priority elections under saturation (#2666)
* Delay voting for non-priority elections under saturation

This is part of https://github.com/nanocurrency/nano-node/pull/2440 . Using a similar strategy of saving the last prioritized difficulty during `update_active_difficulty()`, the result of inserting an election now includes another boolean, hinting that the election may not be a priority due to a large number of active elections and low difficulty of the inserted block.

Active difficulty is now calculated only from these prioritized elections. The cutoff is defined at 10% of thhe `active_elections_size` config which is now 50k. During `request_confirm`, the top elections are prioritized.

When an election is prioritized, a vote is generated for the current winner.

* Make sure to have at least one element when updating last_prioritized_difficulty