Remove assert if a delayed work cache request is not found
In https://github.com/nanocurrency/nano-node/pull/2680 I erased from `delayed_work` in `wallet::action_complete` to prevent doing queueing an outdated request, but this can make it assert when the request is not found later. Removing the assert is enough as the default action is to not do anything about it, which is correct.