Republishing a vote to principal representatives (#2772)
This behavior was recently changed as part of https://github.com/nanocurrency/nano-node/pull/2468 . This specific change has the least impact to bandwidth usage out of all changes in that commit.

The list of peers to which a new vote is republished now includes PRs. Fanout is unchanged at `.5 * sqrt(peers)`. The effective fanout change on the main network is from ~14 to ~17.

This change allows PRs that do not directly connect to other PRs to still see their votes. Ensuring PR connectivity is another approach that will be explored in the future.