Logging exception handler and fix for stable_log_filename on windows (#2880)
* Logging exception handler and fix for stable_log_filename on windows

Although our logger has never thrown in normal conditions, it has been reported to when stressing the RPC in a RAM node, especially with a low rotation file size. This adds an exception handler that handles by simply prints to stderr.

Also fixes an issue with the `stable_log_filename` config on Windows. The `node.log` file was being created with read-only permissions, so writing to the file after restarting the node would throw an exception. However, this can't be unit tested because the file gets the full permissions by default.

The change to `nano::logging::release_file_sink ()` enables the added test.

* Set permissions for pre-existing log file as well, as Boost re-creates with read-only on rotation