Fix inconsistent online representatives list from RPC (#2874)
* Erase rep crawler queried hashes from active recently_confirmed

This minor change increases rep crawler and online reps consistency in small ledgers with low activity - mostly for the beta network.

* Make a backup of online reps to avoid an incorrectly reported list via RPC

The list of online reps is only used externally by RPC, but it has been noted that it is often inconsistent. This is due to clearing the list upon sampling online weight.

With this change, a backup is made on each sample, and RPC will now report the list of both currently sampling reps, and the ones previously sampled.