Transport loopback channel
Adds `nano::transport::loopback_channel` as the initial effort towards removing UDP.

A loopback UDP channel is currently used when processing local votes, and the rep crawler can receive and use that channel if by chance the hash was active. This seems to occur in `node.aggressive_flooding` and sometimes lead to failures. Now, by being explicit about the channel type, these channels are not inserted as peer representatives.

Those making use of an abstract channel to send data must ensure data is not sent to a loopback_channel. This is enforced with a release_assert.