Fix qt wallet hitting a debug_assert when creating blocks
This changes the block creation path of the Qt wallet to use the appropriate difficulty for work generation. By doing so, it fixes the assert reported in #3017.

The normal sending process (clicking Send in the wallet) and receive process (automatically) use the node wallet and thus the appropriate difficulty.

A test is added to make some epoch_2 validation where difficulty is different across block types. There is a similar test for the node internal wallet.