This adds a class called prioritization which tracks accounts to be confirmed in priority order determined by time_since_use * balance order where the time since use is per-account.
The priority order is approximated by segmented each account in to a series of buckets determined by the number of leading 0 bits in the balance. Within each bucket transactions are sorted by last-use timestamp of account where the numerically lowest last-use timestamp represents the account with the oldest modification date and thus, highest priority.