Handle signals with boost asio signal handler
Currently the nano_node has 2 problems:
* it uses the function signal which is not very portable
* it calls possibly blocking code from signal handler context

These problems will become even more crucial when we start using SIGHUP to
reload config. In addition, we plan to introduce more processes and it
makes sense to have a common signal framework for all the processes.

This commit introduces the signal manager, which fixes all the problems
listed above.