Fix core test case election.quorum_minimum_confirm_success (#3242)
The test asks the scheduler to activate an election and then check to see
the election is in the active election container.

However, node.scheduler.activate() is not a blocking operation and the
election activation happens in the background at an unspecified time.

To solve this, the election scheduler can be flushed before checking the
active election container.