Do not repeat core_test in CI
Currently we repeat core_test up to 10 times, in the hope that it succeeds.
But that way, we are hiding problems. We should not hide problems, we
should fix them instead.

I am leaving the loop in the script but simply requesting one iteration
only. If we had a need to repeat tests then it is possible that
we may have that need again someday.