Qt test: disable  wallet.status and introduce wallet status with peer
Create a new test, wallet.status_with_peer, modelled on the test case qt_test wallet.status.

The expected behaviour of the test wallet.status is not defined so it cannot be fixed at the moment and we have bigger problems to work on. See #3235

So, I disable test wallet.status and introduce a new test wallet.status_with_peer, which performs a similar test but with defined behaviour.