Fix for RPC stop command (#3191)
The RPC stop command was not working (the node did exit completely)
because the stop callback was never called because the callback was
called after the node was stopped and the stop callback needed the node
to execute.

This commit removes the node stop before the callback and let's the
stop call back execute, which should stop io_ctx after 3 seconds, which
in turn should exit the daemon and the node should be stopped by its
destructor.