UPnP leases issue (#3298)
* Inverted logic for refreshing upnp port mapping

* Refresh logic for UPnP lease to consider more than one protocol

* mapping_protocol pretty print function to_string()

Preparatory change for better logging in portmapping

* Remove extraneous and needless call to check_mapping_loop

There is no need for the call to check_mapping_loop because there is
an outer check that will work every time as long as it tickets fast enough.

* Remove extraneous and needless logging

* Pretty print function for port_mapping and upnp_state classes

* Check for missing mappings, renewals or gateway changes every health check period

We currently check every health check period for renewals but every
5 minutes for devices. It is too complicated to have 2 different setting
and there is no need. It is harder to code, document and to understand
and there is not much benefit from it.

So, use health check period for all checks. Nice and simple.

* Improve logging regarding remaining upnp lease

The logging before didn't make it clear if it was remaining lease or
total lease time.

* Always print all upnp logs if unpn logging is enabled

Some upnp logs are throttle controlled so that users are not swamped by
upnp messages, if upnp is not working.

However, when upnp logging is explicitly enabled then it makes no sense
to throttle them.

Co-authored-by: Dimitrios Siganos <dimitris@siganos.org>