Fix half principal rep check (#3358)
* Check for half principal even if the account is a duplicate

An account might be added that is not a half principal but might become
a half principal later. So we should check for half principal every time.

* Use have_half_rep() function rather than access variables directly

* Convert half_principal counter to boolean

The counting of half principals without tracking what is counted is
unreliable. Fortunately, we only care about whether a half principal
exists or not. So therefore convert the counter to a boolean so that
the counter does not get used when it is wrong.