Fix for unit test bulk.offline_send
The test expected the ledger to be able to report a block as confirmed
because the active transaction container was empty. But it has to allow
for some time for the data/event to flow to the ledger.

Converted an assert to an assert timely to allow for that propagation.

This is the failure:
[==========] Running 1 test from 1 test suite.
[----------] Global test environment set-up.
[----------] 1 test from bulk
[ RUN      ] bulk.offline_send
/home/ds/nano/branches/develop/nano-node/nano/core_test/bootstrap.cpp:1764: Failure
Value of: node1->block_confirmed (send1->hash ())
  Actual: false
Expected: true
terminate called after throwing an instance of 'testing::internal::GoogleTestFailureException'
  what():  /home/ds/nano/branches/develop/nano-node/nano/core_test/bootstrap.cpp:1764: Failure
Value of: node1->block_confirmed (send1->hash ())
  Actual: false
Expected: true
Aborted