Double unit test timeout in github CI
Sometimes, the core_test and other tests, timeout when running in github
CI. the tell tell sign of a timeout is the return code 124, which is
returned by the timeout command.

The CI computers are slow and as we add more and more tests we need to
adjust this value to accommodate the slowness and the extra tests.

I doubled the timeout so that there is less of a chance of failing
due to timeout.