Remove 1 sec socket timeout in dev mode (becomes 2 sec as in normal mode)
A 1 second timeout seems a little too aggressive even in dev mode.
It can lead to false failures in tests.

Also, it is a complication too far. It is an extra logical branch,
a complication that, as far as I can see, it does not give us anything
useful, except from the possibility of slightly faster unit test runs but
reliability should be more important than a possible slight speed increase
in unit test execution.