Bugfix: node_wrapper using dev net params instead of active (issue #3447)
The node_wrapper was using dev network params instead of active network
params. This caused some CLI commands (at least --wallet_create and
wallet_list) to fail.

To fix the bug, for now, I created a local copy of network_params in
node_wrapper and initialise it with active_network.

However, it would be better to have a network_param reference passed down
so that we do not have multiple copies of network_params. I created an
issue for that improvement:
https://github.com/nanocurrency/nano-node-internalonly/issues/73