Fix handling of SIGSEGV and SIGABRT in nano_node
The handling of SIGSEGV and SIGABRT did not have the desired effect of
generating stack traces of the source code that caused the signal.

That's because the signals were handled by the signal manager, which
first switches stack context to one that is safe to execute any code.
However, for SIGSEGV and SIGABRT we want to print the stacktrace of
the problem and therefore we cannot switch context.

So, I am setting the handling of SIGSEGV and SIGABRT to use the classic
signal function, like before the signal manager was introduced.