Double max_peers_per_ip and max_peers_per_subnet (5 -> 10 and 20 -> 40) (#3589)
We are getting complaints that nodes are hitting this limit. We do not
think it is really a problem but we are increasing this limit to see if
the warnings stop. These complaints could be diverting focus away from
the real problems.

max_peers_per_ip was 5 and now it is 10
max_peers_per_subnet was 20 and now it is 40