Rename NANO_TEST_PEER_NETWORK to NANO_DEFAULT_PEER and apply it to all networks (live, beta, test)
We want to have one environment variable to override the default
preconfigured peer in all networks cases (live, beta and test).

The name thus needs renaming to remove TEST from its name.
Since we are renaming it, we also take the opportunity to improve the name
to something more descriptive, NANO_DEFAULT_PEER.