Fix slow test telemetry.all_peers_use_single_request_cache
And rename it to telemetry.cache_read_and_timeout.

The test was broken because it did a sleep for 7 seconds.
In unit tests, any waiting should be done with ASSERT_TIMELY, or something
similar, so that the system's poll() function is called regularly.

In this case, the poll() function wasn't called for 7 seconds, which
caused things to timeout and comms to be disrupted. The solution was to
replace the sleep with an ASSERT_TIMELY.