Fix for unit test unchecked.multiple_get, do not use store.count()
When using rocksdb as the backend database, the count() method of the
store cannot be trusted, it does not give accurate results.

This problem is fixed by counting the blocks in the unchecked table
one by one, by iterating through all of them, which should work for all
db backends.

resolves #3737