Simplify unit test node.bootstrap_confirm_frontiers
This test failed in CI and I have not identified the reason for the
failure and I cannot recreate the failure.

The test was much more complicated than it needed to be and tested for
intermediate results that maybe do not happen exactly as thought.
I simplified the test, which maybe is a solution. I will assume so for now.

resolves #3771