Removing nano::vote::blocks variant which disables parsing for vote-by-block which is not currently emitted.
The existing implementation was problematic, modeled as a vector-of-variants rather than a variant of a block and hash vector. It also disallowed behavior that can be correctly encoded, such as empty vote-by-hash votes.

Fixing this abstraction inversion was more difficult that simply removing vote by block since this path also introduced its own complexity.