Bugfix: correctly check for magic bytes and network in message header
The network bytes (magic bytes), the first 2 byts of the header, were not
correctly checked. They were checked in the network class inbound function
but that function is called for realtime messages after the handshake is
completed. It is not called for bootstrap messages not handshake messages.

Also the version checking was not done properly so I am doing it now
properly immediately after the magic bytes tests.

The checks are done at the following 3 places:
* nano::bootstrap_server::receive_header_action
* nano::transport::tcp_channels::start_tcp_receive_node_id
* nano::message_parser::deserialize_buffer

The last one is likely only used by UDP and test scripts that use UDP.
So we should be able to remove the last one when we remove the UDP code.

# Conflicts:
#	nano/core_test/network.cpp