Improves the vote processor class and limits its flush() (#142)
- Limit the vote_processor::flush_active()
- The is_active var became useless, also moved the lock position
- Move the condition.notify_all() up
- Improve function flush documentation
- Test normal behavior is getting more votes than processing
- Also updated the invalid_signature test to ASSERT_TIMELY since it
should wait for a flush to happen.
- Also removed the lock for reading total_processed atomic
variable
- Explicitly check for timeout in unit test vote_processor.overflow
- Vote processor: clear votes and notify condition variable on thread
exit

In theory the vote processor can be stopped and have votes left in its
queue which would just be strange and a potential source of problems.

Also, a final kick of the condition variable is needed on thread exit
and after vote clearing.

- Check for confirm req loop blocking too long on vote process flush

This has been a problem in the past and it is always tricky to do flush
operations, so it is prudent to add a check for timeout here and display
a warning in release and throw an assert in debug builds.

- Use std::memory_order_relaxed in vote processor::flush for
total_processed

There is no need to take a memory barrier when reading the atomic
total_processed, so use relaxed memory order.

Co-Authored-By: Dimitrios Siganos <dimitris@siganos.org>

Co-authored-by: Dimitrios Siganos <dimitris@siganos.org>
# Conflicts:
#	nano/core_test/vote_processor.cpp