Fix unit test node.vote_by_hash_republish
The test was failing sometimes because it had a race condition between
delivering the winning forked block and its vote. If the vote arrived
before the block then the test fails because the either the block or the
vote doesn't get republished.

It is not clear if both orders (block before vote, vote before block)
should work. Enforcing the dominant order so that the test passes
consistently. Whether the other order should also pass is treated as a
separate problem.