Fix unit test node, local_votes_cache_generate_new_vote (#3927)
A second node was introduced as part of UDP removal work which is needless
and interferes with the original test case. The second node generates vote
requests (due to rep crawling for instance) and extra votes will be
generated that break this test. The solution is to remove the needless
extra node.