Fix for test election.quorum_minimum_update_weight_before_quorum_checks (#3932)
* Fix for test election.quorum_minimum_update_weight_before_quorum_checks

The problem was that there was a race condition between the rep crawler
background thread and rep_crawler::response(). For the response to be
processed as valid, a request must be marked in rep_crawler::active
container. I added  force flag in rep_crawler::response to insert the vote
regardless to be used in testing.