Fix MacOS compilation, add libsecure to target link libs for nano_lib
The problem appears to be that nano_lib uses libsecure but it is not
listing it as a dependency. The ipc_flatbuffers_test_client does not list
libsecure and the compilation was failing.

The problem was only happening in MacOs Debug builds only.
I assume that it was happening in Debug builds only because of an assert
that used libsecure code.

However, I do not understand why this problem was not appearing on linux.