Move bootstraps from node_flags to node_config
The reason is that we want to be reverse the polarity and the default
values of those values and tehe flags were not able to do that without
adding too many specific command line args.

Also, in the long term, we want to node flags in favour of node config.