Introduce the method systen::start_election()
In many tests, we currently use the method node::block_confirm(), which
uses block processor flush and is therefore broken. It is also terribly
named.

I introduce the method start_election with the aim of replacing
node::block_confirm in unit tests.

I converted a few unit tests to use this new function.