Fix nano::test::blocks_confirm (#4117)
* Fix iace in nano::test::blocks_confirm, use start_election

Refactor nano::test::blocks_confirm to use nano::test::start_election
and fix the race conition that arises when process_active is followed
by blocks_confirm.

* Rename nano::test::blocks_confirm to nano::test::start_elections

* Move nano::test::start_elections to testutil and provide 2 ways to call it

We have many similar functions in testutil.cpp so it makes sense to move
it there. Also, provide a way to call it with hashes only for maximum
reusability.